Parameter Maker.Envelope

type t

The type for envelopes

val dimensions : int

The number of dimensions of this kind of envelope

val compare_dim : int -> t -> t -> int

compare_dim i a b compares a and b along the ith dimension.

Raises Invalid_arg if i >= dimensions.

val t : t Repr.t

A runtime representation of envelopes.

val empty : t

The empty envelope.

val intersects : t -> t -> bool

Whether or not two envelopes intersect.

val merge : t -> t -> t

Computing the envelope that contains both the arguments

val merge_many : t list -> t

Like merge only for many envelopes.

val area : t -> float

Compute the area of an envelope.

val contains : t -> t -> bool

contains a b asks whether b is contained by a.